tests: Migrate PopUpView tests to test_popups.
