+++ tests/ui/test_ui_tools.py
@@ -7,7 +7,7 @@ import pytz
 from bs4 import BeautifulSoup
 from urwid import Columns, Divider, Padding, Text
 
-from zulipterminal.config.keys import is_command_key, keys_for_command
+from zulipterminal.config.keys import keys_for_command
 from zulipterminal.config.symbols import (
     QUOTED_TEXT_MARKER, STREAM_TOPIC_SEPARATOR, TIME_MENTION_MARKER,
 )
@@ -18,8 +18,8 @@ from zulipterminal.ui_tools.buttons import (
 )
 from zulipterminal.ui_tools.views import (
     LeftColumnView, MessageView, MiddleColumnView, ModListWalker,
-    PopUpConfirmationView, PopUpView, RightColumnView, StreamsView,
-    StreamsViewDivider, TopicsView, UsersView,
+    PopUpConfirmationView, RightColumnView, StreamsView, StreamsViewDivider,
+    TopicsView, UsersView,
 )
 
 
@@ -1250,63 +1250,6 @@ class TestLeftColumnView:
                         count=count)
             for topic, count in zip(topic_list, unread_count_list)
         ])
-
-
-class TestPopUpView:
-    @pytest.fixture(autouse=True)
-    def pop_up_view(self, mocker):
-        self.controller = mocker.Mock()
-        mocker.patch.object(self.controller, 'maximum_popup_dimensions',
-                            return_value=(64, 64))
-        self.command = 'COMMAND'
-        self.title = 'Generic title'
-        self.width = 16
-        self.widget = mocker.Mock()
-        mocker.patch.object(self.widget, 'rows', return_value=1)
-        self.widgets = [self.widget, ]
-        self.list_walker = mocker.patch(VIEWS + '.urwid.SimpleFocusListWalker',
-                                        return_value=[])
-        self.super_init = mocker.patch(VIEWS + '.urwid.ListBox.__init__')
-        self.super_keypress = mocker.patch(VIEWS + '.urwid.ListBox.keypress')
-        self.pop_up_view = PopUpView(self.controller, self.widgets,
-                                     self.command, self.width, self.title)
-
-    def test_init(self):
-        assert self.pop_up_view.controller == self.controller
-        assert self.pop_up_view.command == self.command
-        assert self.pop_up_view.title == self.title
-        assert self.pop_up_view.width == self.width
-        self.list_walker.assert_called_once_with(self.widgets)
-        self.super_init.assert_called_once_with(self.pop_up_view.log)
-
-    @pytest.mark.parametrize('key', keys_for_command('GO_BACK'))
-    def test_keypress_GO_BACK(self, key, widget_size):
-        size = widget_size(self.pop_up_view)
-        self.pop_up_view.keypress(size, key)
-        assert self.controller.exit_popup.called
-
-    def test_keypress_command_key(self, mocker, widget_size):
-        size = widget_size(self.pop_up_view)
-        mocker.patch(VIEWS + '.is_command_key', side_effect=(
-            lambda command, key: command == self.command
-        ))
-        self.pop_up_view.keypress(size, 'cmd_key')
-        assert self.controller.exit_popup.called
-
-    def test_keypress_navigation(self, mocker, widget_size,
-                                 navigation_key_expected_key_pair):
-        key, expected_key = navigation_key_expected_key_pair
-        size = widget_size(self.pop_up_view)
-        # Patch `is_command_key` to not raise an 'Invalid Command' exception
-        # when its parameters are (self.command, key) as there is no
-        # self.command='COMMAND' command in keys.py.
-        mocker.patch(VIEWS + '.is_command_key', side_effect=(
-            lambda command, key:
-            False if command == self.command
-            else is_command_key(command, key)
-        ))
-        self.pop_up_view.keypress(size, key)
-        self.super_keypress.assert_called_once_with(size, expected_key)
 
 
 class TestPopUpConfirmationView:
